Converted to using virtualenv
This is a gigantic change, because it now includes an entire
python interpreter and django installation in the folder.

The benefit of this is that it allows us to manage versions
so that we know AWS will work and allows us to have a uniform
development environment.

What this means is that you should issue

    $ source bin/activate
before you begin development. If you want to check that you are
working in a virtualenv, try

    $ which python
and ensure that it looks something like
`/Users/tim/socialmath/bin/python`

If you want to add a pip package, please put it in requirements.txt
with the appropriate version number if applicable.